Reverted the effects protocol
https://github.com/Askannz/msi-perkeyrgb/issues/24
There is a critical issue that LOCKS the RGB controller in an unusable
state, which persists after reboot. At least 2 machines are affected,
including mine. Still trying to figure out a solution to reset the
controller and make it usable again, bit in the meantime I'm reverting
all the new features that may have caused this.